Target Rust stable channel
This commit adds `/rust-toolchain` to the root repository directory.
This will (hopefully) enforce (depending if rustup is used) the stable
channel to be used for the project.

The reason I'm now targeting the Rust stable channel is to ensure
stability, and that scrobblers doesn't use nightly features at this
moment in time.